+++ src/main/java/blackjack/view/dto/PlayerEarning.java
@@ -1,4 +1,4 @@
 package blackjack.view.dto;
 
-public record PlayerEarning(String name, double earning) {
+public record PlayerEarning(String name, int earning) {
 }
